refactor!: revamped project structure
