--- global.go
@@ -1,5 +1,0 @@
-package main
-
-var CurrentConfig Config
-var CurrentSong SongData = SongData{LyricsType: 5}
-var IsPlaying bool
