+++ internal/loop/syncloop.go
@@ -1,13 +1,18 @@
-package main
+package loop
 
 import (
 	"fmt"
 	"math"
 	"time"
+
+	"lrcsnc/internal/lrclib"
+	"lrcsnc/internal/player"
+	"lrcsnc/internal/romanization"
+	"lrcsnc/pkg"
 )
 
 func SyncLoop() {
-	checkerTicker := time.NewTicker(time.Duration(CurrentConfig.Playerctl.PlayerctlSongCheckInterval*1000) * time.Millisecond)
+	checkerTicker := time.NewTicker(time.Duration(pkg.CurrentConfig.Playerctl.PlayerctlSongCheckInterval*1000) * time.Millisecond)
 	positionCheckTimer := time.NewTimer(time.Second)
 	positionInnerCheckTicker := time.NewTicker(time.Second)
 	positionInnerCheckTicker.Stop()
@@ -22,11 +27,11 @@ func SyncLoop() {
 	go func() {
 		for {
 			<-checkerTicker.C
-			song := GetSongData()
-			if song.Song != CurrentSong.Song || song.Artist != CurrentSong.Artist || song.Album != CurrentSong.Album || song.Duration != CurrentSong.Duration {
-				_, position = GetPlayerData()
+			song := player.GetSongData()
+			if song.Song != pkg.CurrentSong.Song || song.Artist != pkg.CurrentSong.Artist || song.Album != pkg.CurrentSong.Album || song.Duration != pkg.CurrentSong.Duration {
+				_, position = player.GetPlayerData()
 				timeBeforeGettingLyrics = time.Now()
-				CurrentSong = song
+				pkg.CurrentSong = song
 
 				songChanged <- true
 			}
@@ -39,15 +44,15 @@ func SyncLoop() {
 			<-songChanged
 
 			// If the duration equals 0s, then there are no supported players out there.
-			if CurrentSong.Duration == 0 {
-				CurrentSong.LyricsType = 4
+			if pkg.CurrentSong.Duration == 0 {
+				pkg.CurrentSong.LyricsType = 4
 				fullLyrChan <- false
 				continue
 			}
 
-			GetSyncedLyrics(&CurrentSong)
-			if CurrentSong.LyricsType == 5 {
-				CurrentSong.LyricsType = 6
+			lrclib.GetSyncedLyrics(&pkg.CurrentSong)
+			if pkg.CurrentSong.LyricsType == 5 {
+				pkg.CurrentSong.LyricsType = 6
 			}
 
 			fullLyrChan <- true
@@ -60,13 +65,13 @@ func SyncLoop() {
 		for {
 			<-positionCheckTimer.C
 			var initialPosition float64
-			IsPlaying, initialPosition = GetPlayerData()
-			if IsPlaying {
-				if CurrentConfig.Global.DisableActiveSync {
+			pkg.IsPlaying, initialPosition = player.GetPlayerData()
+			if pkg.IsPlaying {
+				if pkg.CurrentConfig.Global.DisableActiveSync {
 					positionInnerCheckTicker.Reset(time.Second)
 					<-positionInnerCheckTicker.C
 					var newPosition float64
-					IsPlaying, newPosition = GetPlayerData()
+					pkg.IsPlaying, newPosition = player.GetPlayerData()
 					diff := newPosition - initialPosition
 					if !(diff >= 0.9 && diff <= 1.1) {
 						UpdatePosition(newPosition)
@@ -77,10 +82,10 @@ func SyncLoop() {
 					for i := 0; i < requiredTicks; i++ {
 						<-positionInnerCheckTicker.C
 						var newPosition float64
-						IsPlaying, newPosition = GetPlayerData()
+						pkg.IsPlaying, newPosition = player.GetPlayerData()
 						expectedPosition := (initialPosition + 0.1*(float64(i)+1))
 						diff := newPosition - expectedPosition
-						if !(((diff <= 0.21 && diff >= -1.11) || (diff >= 0.89 && diff <= 1.01)) || !IsPlaying) {
+						if !(((diff <= 0.21 && diff >= -1.11) || (diff >= 0.89 && diff <= 1.01)) || !pkg.IsPlaying) {
 							UpdatePosition(newPosition)
 							break
 						}
@@ -101,11 +106,11 @@ func SyncLoop() {
 
 			prevLyric := ""
 			count := 1
-			for i, lyric := range CurrentSong.Lyrics {
-				if CurrentConfig.Output.Romanization.IsEnabled() && IsSupportedAsianLang(lyric) {
-					CurrentSong.Lyrics[i] = Romanize(lyric)
+			for i, lyric := range pkg.CurrentSong.Lyrics {
+				if pkg.CurrentConfig.Output.Romanization.IsEnabled() && romanization.IsSupportedAsianLang(lyric) {
+					pkg.CurrentSong.Lyrics[i] = romanization.Romanize(lyric)
 				}
-				if CurrentConfig.Output.ShowRepeatedLyricsMultiplier {
+				if pkg.CurrentConfig.Output.ShowRepeatedLyricsMultiplier {
 					if lyric == prevLyric && lyric != "" {
 						count++
 					} else {
@@ -114,18 +119,18 @@ func SyncLoop() {
 					}
 
 					if count != 1 {
-						if CurrentConfig.Output.PrintRepeatedLyricsMultiplierToTheRight {
-							CurrentSong.Lyrics[i] = fmt.Sprintf(lyric+" "+CurrentConfig.Output.RepeatedLyricsMultiplierFormat, count)
+						if pkg.CurrentConfig.Output.PrintRepeatedLyricsMultiplierToTheRight {
+							pkg.CurrentSong.Lyrics[i] = fmt.Sprintf(lyric+" "+pkg.CurrentConfig.Output.RepeatedLyricsMultiplierFormat, count)
 						} else {
-							CurrentSong.Lyrics[i] = fmt.Sprintf(CurrentConfig.Output.RepeatedLyricsMultiplierFormat+" "+lyric, count)
+							pkg.CurrentSong.Lyrics[i] = fmt.Sprintf(pkg.CurrentConfig.Output.RepeatedLyricsMultiplierFormat+" "+lyric, count)
 						}
 					}
 				}
 			}
 
-			if CurrentConfig.Output.TimestampOffset != 0 {
-				for i, timestamp := range CurrentSong.LyricTimestamps {
-					CurrentSong.LyricTimestamps[i] = timestamp + (float64(CurrentConfig.Output.TimestampOffset) / 1000)
+			if pkg.CurrentConfig.Output.TimestampOffset != 0 {
+				for i, timestamp := range pkg.CurrentSong.LyricTimestamps {
+					pkg.CurrentSong.LyricTimestamps[i] = timestamp + (float64(pkg.CurrentConfig.Output.TimestampOffset) / 1000)
 				}
 			}
