+++ internal/lrclib/get.go
@@ -1,4 +1,4 @@
-package main
+package lrclib
 
 import (
 	"encoding/json"
@@ -8,14 +8,17 @@ import (
 	"math"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
+
+	"lrcsnc/internal/cache"
+	"lrcsnc/pkg"
+	"lrcsnc/util"
 )
 
 // Sets the Lyrics and LyricTimestamps properties in SongData object.
-func GetSyncedLyrics(song *SongData) {
-	if CurrentConfig.Cache.Enabled {
-		cachedData, isNotExpired := GetCachedLyrics(song)
+func GetSyncedLyrics(song *pkg.SongData) {
+	if pkg.CurrentConfig.Cache.Enabled {
+		cachedData, isNotExpired := cache.GetCachedLyrics(song)
 		if isNotExpired && !(len(cachedData.Lyrics) == 0 && !cachedData.Instrumental) {
 			song.LyricTimestamps = cachedData.LyricTimestamps
 			song.Lyrics = cachedData.Lyrics
@@ -65,7 +68,7 @@ func GetSyncedLyrics(song *SongData) {
 
 		for i, lyric := range syncedLyrics {
 			lyricParts := strings.SplitN(lyric, " ", 2)
-			timecode := timecodeStrToFloat(lyricParts[0])
+			timecode := util.TimecodeToFloat(lyricParts[0])
 			if timecode == -1 {
 				continue
 			}
@@ -83,8 +86,8 @@ func GetSyncedLyrics(song *SongData) {
 		song.LyricTimestamps = resultTimestamps
 	}
 
-	if CurrentConfig.Cache.Enabled && song.LyricsType != 1 {
-		var dataToBeCached Cache = Cache{
+	if pkg.CurrentConfig.Cache.Enabled && song.LyricsType != 1 {
+		var dataToBeCached cache.Cache = cache.Cache{
 			LyricTimestamps: resultTimestamps,
 			Lyrics:          resultLyrics,
 			Instrumental:    false,
@@ -92,30 +95,14 @@ func GetSyncedLyrics(song *SongData) {
 		if song.LyricsType == 2 {
 			dataToBeCached.Instrumental = true
 		}
-		if err := StoreCachedLyrics(*song, dataToBeCached); err != nil {
+		if err := cache.StoreCachedLyrics(*song, dataToBeCached); err != nil {
 			log.Println("Could not save the lyrics to the cache! Are there writing perms?")
 		}
 	}
 }
 
-func timecodeStrToFloat(timecode string) float64 {
-	// [00:00.00]
-	if len(timecode) != 10 {
-		return -1
-	}
-	minutes, err := strconv.ParseFloat(timecode[1:3], 64)
-	if err != nil {
-		return -1
-	}
-	seconds, err := strconv.ParseFloat(timecode[4:9], 64)
-	if err != nil {
-		return -1
-	}
-	return minutes*60.0 + seconds
-}
-
 // Make a URL to lrclib.net/api/get to send a GET request to
-func makeURLGet(song *SongData) url.URL {
+func makeURLGet(song *pkg.SongData) url.URL {
 	lrclibURL, err := url.Parse("http://lrclib.net/api/get?" + url.PathEscape(fmt.Sprintf("track_name=%v&artist_name=%v&album_name=%v&duration=%v", song.Song, song.Artist, song.Album, int(math.Ceil(song.Duration)))))
 	if err != nil {
 		log.Fatalln(err)
@@ -124,7 +111,7 @@ func makeURLGet(song *SongData) url.URL {
 }
 
 // Make a URL to lrclib.net/api/search with album data to send a GET request to
-func makeURLSearchWithAlbum(song *SongData) url.URL {
+func makeURLSearchWithAlbum(song *pkg.SongData) url.URL {
 	lrclibURL, err := url.Parse("http://lrclib.net/api/search?" + url.PathEscape(fmt.Sprintf("track_name=%v&artist_name=%v&album_name=%v", song.Song, song.Artist, song.Album)))
 	if err != nil {
 		log.Fatalln(err)
@@ -133,7 +120,7 @@ func makeURLSearchWithAlbum(song *SongData) url.URL {
 }
 
 // Make a URL to lrclib.net/api/search only with necessary data (song name and artist name) to send a GET request to
-func makeURLSearch(song *SongData) url.URL {
+func makeURLSearch(song *pkg.SongData) url.URL {
 	lrclibURL, err := url.Parse("http://lrclib.net/api/search?" + url.PathEscape(fmt.Sprintf("track_name=%v&artist_name=%v", song.Song, song.Artist)))
 	if err != nil {
 		log.Fatalln(err)
