+++ internal/lrclib/structs.go
@@ -1,0 +1,7 @@
+package lrclib
+
+type LrcLibJson struct {
+	Instrumental bool   `json:"instrumental"`
+	PlainLyrics  string `json:"plainLyrics"`
+	SyncedLyrics string `json:"syncedLyrics"`
+}
