+++ pkg/isPlaying.go
@@ -1,0 +1,3 @@
+package pkg
+
+var IsPlaying bool
