+++ pkg/songdata.go
@@ -1,0 +1,3 @@
+package pkg
+
+var CurrentSong SongData = SongData{LyricsType: 5}
