+++ util/removeBadCharacters.go
@@ -1,0 +1,7 @@
+package util
+
+import "regexp"
+
+func RemoveBadCharacters(str string) string {
+	return regexp.MustCompile(`[:;!?,\.\[\]<>\/\\*|]+`).ReplaceAllString(str, "_")
+}
